refactor: extract src/shared/hooks/ and src/shared/ui/ to standalone directories
Move React runtime code out of src/shared/ so it remains a pure
types+utils boundary:

- src/shared/hooks/ → src/hooks/
- src/shared/ui/ → src/ui/
- TypedClientSocket/ConnectionStatus types → src/shared/socketEvents.ts
- Remove src/shared/hooks/** and src/shared/ui/** ESLint ignores
- Update all import paths across 20+ consumer files
- Update active documentation references

Closes #105
