+++ src/combat/KonvaMap.tsx
@@ -13,7 +13,7 @@ import { TokenTooltip } from './TokenTooltip'
 import { MeasureTool } from './tools/MeasureTool'
 import { RangeTemplate } from './tools/RangeTemplate'
 import { snapToGrid } from './combatUtils'
-import { useToast } from '../shared/ui/useToast'
+import { useToast } from '../ui/useToast'
 import { useCameraControls } from './hooks/useCameraControls'
 import { useTokenAwareness } from './hooks/useTokenAwareness'
 import { useEntityDrop } from './hooks/useEntityDrop'
