+++ src/ui/ResourceBar.tsx
